+++ projects/main/src/main/java/libshapedraw/LibShapeDraw.java
@@ -10,22 +10,21 @@ import libshapedraw.internal.LSDInternalException;
 import libshapedraw.shape.Shape;
 
 /**
- * Main API entry point, instantiated by client code.
- * Multiple API instances can coexist, each with their own settings and state.
+ * Main entry point for the LibShapeDraw API.
  * <p>
- * Instantiating this class will automatically register it with the internal
- * Controller. Any shapes you add with be rendered and any listeners you add
+ * Instantiating this class will automatically register it. Shapes added using
+ * addShape will be rendered. LSDEventListeners added using addEventListener
  * will receive events.
  * <p>
- * API instances are not thread-safe. Accessing or modifying getShapes() (or
- * any other exposed collection) from any thread other than the main Minecraft
- * game thread may result in non-deterministic behavior.
+ * Multiple API instances can coexist, each with their own settings and state.
  * <p>
- * @see the demos in src/test/java for sample usage
+ * See the demos in projects/demos/src/main/java for sample usage.
  */
 public class LibShapeDraw {
     private final Set<Shape> shapes;
+    private final Set<Shape> shapesReadonly;
     private final Set<LSDEventListener> eventListeners;
+    private final Set<LSDEventListener> eventListenersReadonly;
     private final String instanceId;
     private boolean visible = true;
     private boolean visibleWhenHidingGui = false;
@@ -33,17 +32,26 @@ public class LibShapeDraw {
     public LibShapeDraw() {
         this(Thread.currentThread().getStackTrace()[2].toString());
     }
+
+    /**
+     * @param ownerId optional identifier for the mod that is using this API
+     *                instance.
+     */
     public LibShapeDraw(String ownerId) {
         shapes = Collections.checkedSet(new LinkedHashSet<Shape>(), Shape.class);
+        shapesReadonly = Collections.unmodifiableSet(shapes);
         eventListeners = Collections.checkedSet(new LinkedHashSet<LSDEventListener>(), LSDEventListener.class);
+        eventListenersReadonly = Collections.unmodifiableSet(eventListeners);
         instanceId = LSDController.getInstance().registerApiInstance(this, ownerId);
     }
 
     /**
-     * Permanently unregister from the internal Controller. Normally there's no
-     * need to ever do this; it generally makes more sense to call
-     * setVisible(false), remove event listeners, and/or clear the Shapes
-     * collection if you want to temporarily shut things down.
+     * Permanently unregister the entire API instance. It will no longer render
+     * its Shapes or dispatch LSDEvents to registered listeners.
+     * <p>
+     * Normally there's no need to ever do this. It generally makes more sense
+     * to call setVisible(false) or clearShapes() to temporarily shut things
+     * down.
      */
     public boolean unregister() {
         return LSDController.getInstance().unregisterApiInstance(this);
@@ -96,9 +104,16 @@ public class LibShapeDraw {
     public boolean isVisible() {
         return visible;
     }
+
     /**
      * If set to false, none of the shapes owned by this API instance will be
      * rendered. Defaults to true.
+     * <p>
+     * When set to false, the shapes still exist. They just won't be rendered
+     * until this property is set to true again.
+     * <p>
+     * To show/hide individual shapes rather than the entire shape collection,
+     * use {@link Shape#setVisible}.
      */
     public LibShapeDraw setVisible(boolean visible) {
         this.visible = visible;
@@ -109,12 +124,15 @@ public class LibShapeDraw {
     public boolean isVisibleWhenHidingGui() {
         return visibleWhenHidingGui;
     }
+
     /**
-     * If set to true, the shapes owned by this API instance are treated as
-     * in-game objects, rendered regardless of whether the GUI is visible.
+     * If set to true, the shapes owned by this API instance are rendered
+     * regardless of whether the GUI is visible. In other words, the shapes are
+     * considered part of the game world rather than part of the GUI.
      * <p>
-     * If set to false (default), the shapes are treated as part of the GUI and
-     * will be appropriately hidden when the user presses F1 to hide the GUI.
+     * If set to false (default), the shapes are considered part of the GUI.
+     * They will be appropriately hidden when the user presses F1 to hide the
+     * GUI.
      * <p>
      * @see {@link #setVisible}, which overrides this property when set.
      */
@@ -123,49 +141,134 @@ public class LibShapeDraw {
         return this;
     }
 
-    /** This is not thread-safe. */
-    public Set<Shape> getShapes() {
-        return shapes;
-    }
     /**
-     * Convenience method, equivalent to getShapes().add(shape).
-     * This is not thread-safe.
+     * Get a read-only view of the full set of shapes registered to this API
+     * instance. To modify this set use addShape, removeShape, and clearShapes.
+     */
+    public Set<Shape> getShapes() {
+        return shapesReadonly;
+    }
+
+    /**
+     * Register a Shape to be rendered by this API instance.
+     * <p>
+     * Thread safety is not guaranteed. To avoid non-deterministic behavior,
+     * only call this method from the main Minecraft thread.
      * @returns the instance (for method chaining)
      */
     public LibShapeDraw addShape(Shape shape) {
-        getShapes().add(shape);
-        return this;
-    }
-    /**
-     * Convenience method, equivalent to getShapes().remove(shape).
-     * This is not thread-safe.
-     * @returns the instance (for method chaining)
-     */
-    public LibShapeDraw removeShape(Shape shape) {
-        getShapes().remove(shape);
+        if (shape == null) {
+            throw new IllegalArgumentException();
+        }
+        shapes.add(shape);
+        shape.onAdd(this);
         return this;
     }
 
-    /** This is not thread-safe. */
-    public Set<LSDEventListener> getEventListeners() {
-        return eventListeners;
-    }
     /**
-     * Convenience method, equivalent to getEventListeners().add(listener)
-     * This is not thread-safe.
+     * Unregister a Shape, no longer rendering it.
+     * <p>
+     * Attempting to remove a shape that is not part of this API instance's
+     * shape collection is allowed but won't do anything.
+     * <p>
+     * Thread safety is not guaranteed. To avoid non-deterministic behavior,
+     * only call this method from the main Minecraft thread.
+     * 
+     * @see {@link Shape#setVisible} for an alternate way of preventing a shape
+     *      from rendering. Generally, removeShape should be called if the
+     *      shape will never be rendered again. If the shape just needs to be
+     *      hidden temporarily, use setVisible.
+     * 
+     * @returns the instance (for method chaining)
+     */
+    public LibShapeDraw removeShape(Shape shape) {
+        if (shapes.remove(shape)) {
+            shape.onRemove(this);
+        }
+        return this;
+    }
+
+    /**
+     * Unregister all Shapes owned by this API instance, no longer rendering
+     * them.
+     * <p>
+     * Thread safety is not guaranteed. To avoid non-deterministic behavior,
+     * only call this method from the main Minecraft thread.
+     * 
+     * @see {@link setVisible} for an alternate way of preventing all shapes
+     *      from rendering. Generally, clearShapes should be called if the
+     *      current shapes will never be rendered again. If the shapes just
+     *      need to be hidden temporarily, use setVisible.
+     * 
+     * @returns the instance (for method chaining)
+     */
+    public LibShapeDraw clearShapes() {
+        LinkedHashSet<Shape> prev = new LinkedHashSet<Shape>(shapes);
+        shapes.clear();
+        for (Shape shape : prev) {
+            shape.onRemove(this);
+        }
+        return this;
+    }
+
+    /**
+     * Get a read-only view of the set of event listeners associated with this
+     * API instance. To modify this set use addEventListener,
+     * removeEventListener, and clearEventListeners.
+     * <p>
+     * Thread safety is not guaranteed. To avoid non-deterministic behavior,
+     * only call this method from the main Minecraft thread.
+     */
+    public Set<LSDEventListener> getEventListeners() {
+        return eventListenersReadonly;
+    }
+
+    /**
+     * Register an event listener to receive LSDEvents.
+     * <p>
+     * Thread safety is not guaranteed. To avoid non-deterministic behavior,
+     * only call this method from the main Minecraft thread.
      * @returns the instance (for method chaining)
      */
     public LibShapeDraw addEventListener(LSDEventListener listener) {
-        getEventListeners().add(listener);
+        if (listener == null) {
+            throw new IllegalArgumentException();
+        }
+        eventListeners.add(listener);
         return this;
     }
+
     /**
-     * Convenience method, equivalent to getEventListeners().remove(listener)
-     * This is not thread-safe.
+     * Unregister an event listener. It will no longer receive LSDEvents.
+     * <p>
+     * Normally there is no need to manually unregister event listeners, but
+     * this method is here if needed.
+     * <p>
+     * Attempting to remove an event listener that is not registered to this
+     * API instance is allowed but won't do anything.
+     * <p>
+     * Thread safety is not guaranteed. To avoid non-deterministic behavior,
+     * only call this method from the main Minecraft thread.
      * @returns the instance (for method chaining)
      */
     public LibShapeDraw removeEventListener(LSDEventListener listener) {
-        getEventListeners().remove(listener);
+        eventListeners.remove(listener);
+        return this;
+    }
+
+    /**
+     * Unregister all event listeners that were registered to this API
+     * instance.
+     * <p>
+     * Normally there is no need to manually unregister event listeners, but
+     * this method is here if needed.
+     * <p>
+     * Thread safety is not guaranteed. To avoid non-deterministic behavior,
+     * only call this method from the main Minecraft thread.
+     * @returns the instance (for method chaining)
+     */
+    public LibShapeDraw clearEventListeners() {
+        eventListeners.clear();
         return this;
     }